+++ src/main/java/tornadofx/CSS.kt
@@ -289,6 +289,7 @@ open class Stylesheet(vararg val imports: KClass<out Stylesheet>) : SelectionHol
         val previousMonth by cssclass()
         val progress by cssclass()
         val progressBar by cssclass()
+        val progressBarTableCell by cssclass()
         val progressIndicator by cssclass()
         val radio by cssclass()
         val radioButton by cssclass()